Update Controller Service and Repository Layer for the feature info

+++ src/main/java/com/up42/repo/FeaturesRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 /**
  * @author jalajchawla
  */
-public interface FeaturesRepository extends JpaRepository<Features,Integer> {
+public interface FeaturesRepository extends JpaRepository<Features,Long> {
 }
